change logic to include all submodules and exclude parent modules (they don't produce JARs)

+++ .github/workflows/scripts/.pinot_pr_comparer.sh
@@ -20,20 +20,24 @@ latest=$(echo "$prnums" | jq '.[0].number')
 gh pr checkout "$latest"
 
 version="$(mvn help:evaluate -Dexpression=project.version -q -DforceStdout | tr -d "%")" #there's a % at the end for some reason
-modnames="$(mvn -pl :pinot help:effective-pom | grep "<module>" | tr -d "/<>" | tr "\n" " ")"
+modnames="$(mvn -pl :pinot help:effective-pom -amd | grep "<module>" | tr -d "/<>" | tr "\n" " ")"
 modnames=${modnames//"module"} # removes the word 'module' from the output
 IFS=' ' read -r -a namelist <<< "$modnames"
 
 mvn clean install -DskipTests
 for name in "${namelist[@]}"; do # eventually remove temp and switch to namelist directly
-  mv "$name"/target/"$name"-"$version".jar commit_jars_new
+  if [ -f "$name"/target/"$name"-"$version".jar ]; then
+    mv "$name"/target/"$name"-"$version".jar commit_jars_new
+  fi
 done
 
 sndlatest=$(echo "$prnums" | jq '.[1].number')
 gh pr checkout "$sndlatest"
 mvn clean install -DskipTests
 for name in "${namelist[@]}"; do # eventually remove temp and switch to namelist directly
-  mv "$name"/target/"$name"-"$version".jar commit_jars_old
+  if [ -f "$name"/target/"$name"-"$version".jar ]; then
+      mv "$name"/target/"$name"-"$version".jar commit_jars_old
+    fi
 done
 
 #eventually change the below to just checking out the gh-pages branch
@@ -57,13 +61,14 @@ if [ -e japicmp_test.txt ]; then
 else
   touch japicmp_test.txt
 fi
-for name in "${namelist[@]}"; do
-  if [ ! -e commit_jars_old/"$name"-"$version".jar ] || [ ! -e commit_jars_new/"$name"-"$version".jar ]; then
-    echo "It seems ${name} does not exist in one or both of the pull requests. Please look into this." >> japicmp_test.txt
+for filename in commit_jars_new/*; do
+  if [ ! -e commit_jars_old/filename ]; then
+    echo "It seems ${name} does not exist in the previous pull request. Please make sure this is intended." >> japicmp_test.txt
+    echo "" >> japicmp_test.txt
     continue
   fi
-  OLD=commit_jars_old/"${name}"-"$version".jar
-  NEW=commit_jars_new/"${name}"-"$version".jar
+  OLD=commit_jars_old/filename
+  NEW=commit_jars_new/filename
   java -jar japicmp.jar \
     --old "$OLD" \
     --new "$NEW" \
